mypy

+++ allennlp/common/tqdm.py
@@ -4,10 +4,10 @@ This wraps tqdm so we can add configurable global defaults for certain tqdm para
 from tqdm import tqdm as _tqdm
 
 class Tqdm:
-    default_mininterval: int = 0.1
+    default_mininterval: float = 0.1
 
     @staticmethod
-    def set_default_mininterval(value: int) -> None:
+    def set_default_mininterval(value: float) -> None:
         Tqdm.default_mininterval = value
 
     @staticmethod
